+++ library/FieldUtils/ProcessModules/ProcessBodyFittedVelocity.cpp
@@ -865,7 +865,7 @@ void ProcessBodyFittedVelocity::GetVelAndConvertToCartSys(
                 if (m_f->m_exp[0]->GetWaveSpace())
                 {
                     m_f->m_exp[0]->HomogeneousBwdTrans(
-                       npoints, m_f->m_exp[var_offset + i]->GetPhys(), vel[i]);
+                        npoints, m_f->m_exp[var_offset + i]->GetPhys(), vel[i]);
                 }
                 else
                 {