+++ solvers/VortexWaveInteraction/utilities/MeshMove.cpp
@@ -1887,8 +1887,7 @@ int main(int argc, char *argv[])
             ASSERTL0(Vmath::Vmax(nVertTot, xold, 1) ==
                          Vmath::Vmax(np_lay, layers_x[m], 1),
                      "  different layer xmax val");
-
-        } 
+        }
 
         // MoveOutsidePointsfixedxpos(npedge, graphShPt,xold_c, yold_c,
         // xold_low, yold_low, 	         xold_up, yold_up, layers_y[0],