+++ CHANGELOG.md
@@ -43,6 +43,7 @@ v5.3.0
 - Fix typos in Vmath and VDmath (!1480)
 - Fix minor typo and removed unused functions in LibUtilities/TimeIntegration (!1476) 
 - Fix RK5 time integration scheme (!1482)
+- Fix ESDIRK time integration scheme (!1484)
 
 **Python**
 - Add wrappers for Interpreter and Equation classes (!1329)